Fixes tests on Python 3.10
Some of the tests were not passing on 3.10. This is due to astropy.
astropy 7 is not supported on 3.10, and Table in astropy 6 dies on
initialization when the constructor is passed name/dtype/units arguments
with an empty list (for rows). Solution I found is to initialize with
rows point to empty list themselves. Sucks, but it seems to work.